pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "CosQA"
version = "0.1.0"
description = "Code repository Q&A pair generator"
authors = [
    {name = "Xinyun Zhang", email = "xinyunz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "fsspec==2024.10.0",
    "instructor==1.5.2",
    "jiter==0.5.0",
    "litellm==1.59.7",
    "openai==1.60.1",
    "packaging==24.2",
    "rapidfuzz==3.11.0",
    "rich==13.9.4",
    "shellingham==1.5.4",
    "typer==0.15.1",
    "python-dotenv==1.0.1",
    "tqdm==4.67.1",
    "pydantic==2.10.6",
    "requests==2.32.3",
    "networkx==3.4.2",
    "pandas==2.2.3",
    "numpy==1.26.4",
    "transformers==4.49.0",
    "torch==2.6.0",
    "sentence-transformers==3.4.1",
    "datasets==3.4.1",
    "huggingface-hub==0.27.1",
    "unsloth",
    "anthropic==0.38.0",
    "llama-index==0.10.68",
    "llama-index-core==0.10.68.post1",
    "scikit-learn==1.6.1",
    "streamlit==1.41.1",
    "tiktoken==0.7.0",
    "pillow==11.0.0",
    "matplotlib==3.9.2",
    "plotly==5.24.1",
    "nltk==3.9.1",
    "beautifulsoup4==4.12.3",
    "faiss-cpu==1.9.0.post1",
    "pypdf==4.3.1",
    "SQLAlchemy==2.0.37",
    "tree-sitter==0.24.0",
    "tree-sitter-java==0.23.5",
    "tree-sitter-python==0.23.6",
    "llama-index-embeddings-voyageai==0.1.4",
    "seaborn==0.13.2",
    "ijson==3.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "black>=21.5b2",
    "isort==5.13.2",
    "flake8>=3.9.2",
]

[project.scripts]
generate-questions = "repo_qa_generator.cli:main"

[tool.setuptools]
packages = ["repo_qa_generator"]